+++ actions/src/robocorp/actions/_args_dispatcher.py
@@ -1,8 +1,10 @@
 import sys
 import typing
 from contextlib import contextmanager
+from typing import Optional
 
 from robocorp.tasks._argdispatch import _ArgDispatcher
+from robocorp.tasks._customization._plugin_manager import PluginManager
 
 
 def _translate(msg):
@@ -14,10 +16,11 @@ def _translate(msg):
 
 
 class _ActionsArgDispatcher(_ArgDispatcher):
-    def __init__(self, *args, **kwargs):
+    def __init__(self, *args, **kwargs) -> None:
         super().__init__(*args, **kwargs)
         self.register("list")(self._list)
         self.register("run")(self._run)
+        self._pm: Optional[PluginManager] = None
 
     @contextmanager
     def _register_lint(self, stream: typing.IO):
@@ -45,7 +48,9 @@ class _ActionsArgDispatcher(_ArgDispatcher):
                 files_found.add(filename)
                 errors = list(
                     x.to_lsp_diagnostic()
-                    for x in _lint_action.iter_lint_errors(filename.read_bytes())
+                    for x in _lint_action.iter_lint_errors(
+                        filename.read_bytes(), self._pm
+                    )
                 )
                 if errors:
                     found_critical = False
@@ -140,7 +145,7 @@ class _ActionsArgDispatcher(_ArgDispatcher):
 
         return ArgumentParserTranslated
 
-    def _dispatch(self, parsed) -> int:
+    def _dispatch(self, parsed, pm: Optional[PluginManager] = None) -> int:
         # Custom dispatch as we need to account for custom flags.
         if not parsed.command:
             self._create_argparser().print_help()
@@ -149,5 +154,7 @@ class _ActionsArgDispatcher(_ArgDispatcher):
         method = self._name_to_func[parsed.command]
         dct = parsed.__dict__.copy()
         dct.pop("command")
+        dct["pm"] = pm
+        self._pm = pm
 
         return method(**dct)