+++ browser/src/robocorp/browser/__init__.py
@@ -19,7 +19,7 @@ from playwright.sync_api import (
 
 from ._types import BrowserEngine, BrowserNotFound, InstallError
 
-__version__ = "2.3.0"
+__version__ = "2.3.1"
 version_info = [int(x) for x in __version__.split(".")]