+++ tasks/src/robocorp/tasks/_constants.py
@@ -1,4 +1,9 @@
-# TODO: Support array and object.
+# Note: besides the types supported here, pydantic is also supported
+# by checking duck-typing for the APIs below:
+#
+# `cls.model_validate(dict)`
+# `cls.model_json_schema()`
+# `obj.model_dump_json()`
 SUPPORTED_TYPES_IN_SCHEMA = (str, int, float, bool)
 
 DEFAULT_TASK_SEARCH_GLOB = "*task*.py"